requirements.txt
neat-python==0.92
numpy==1.26.2
pygame==2.5.2
